fix: change data type in notifications table

+++ database/migrations/2023_07_02_145957_create_notifications_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
 			->onDelete('cascade');
 			$table->boolean('seen');
 			$table->string('type');
-			$table->unsignedTinyInteger('post_author');
+			$table->unsignedBigInteger('post_author');
 			$table->timestamps();
 		});
 	}
